requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.23.2
